first step to centralized settings management

+++ src/Command/Base.php
@@ -54,7 +54,15 @@ abstract class Base implements Handler
      */
     public static function provide(Client $client, Session $session, $config = array(), LoggerInterface $logger = null)
     {
-        return new static($client, $session, $config, $logger);
+        $handler = new static($client, $session, $config, $logger);
+
+        // init settings if any
+        if (isset($config['config_bot']['settings'])) {
+            $settingsInitClass = $config['config_bot']['settings_namespace'].'Initialize';
+            $settingsInitClass::init($handler);
+        }
+
+        return $handler;
     }
 
     /**
@@ -64,37 +72,59 @@ abstract class Base implements Handler
         return get_called_class();
     }
 
-//    protected function initSession($session_id)
-//    {
-//        session_id($session_id);
-//        session_start();
-//        $this->session = $_SESSION;
-//        return $this->session;
-//    }
-//
-//    /**
-//     * @return mixed
-//     */
-//    protected function getSession()
-//    {
-//        return $this->session;
-//    }
+    /**
+     * @return array
+     */
+    public function getConfig()
+    {
+        return $this->config;
+    }
 
     /**
-     * @return Session
+     * @return LoggerInterface
      */
-//    public function getSession()
-//    {
-//        return $this->session;
-//    }
+    public function getLogger()
+    {
+        return $this->logger;
+    }
+
+    /**
+     * @param LoggerInterface $logger
+     * @return Base
+     */
+    public function setLogger($logger)
+    {
+        $this->logger = $logger;
+        return $this;
+    }
+
+    /**
+     * Returns chat_id derived from session
+     * (it must always be set, otherwise an exception is thrown during session initialization)
+     * @return string
+     */
+    public function getChatId()
+    {
+        return $this->session->getSessionId();
+    }
 
     protected function setState($var, $value)
     {
+        // protected states..
+        if (in_array($var, ['settings'])) {
+            return;
+        }
+
         $this->session->set($var, $value);
     }
 
     protected function unsetState($var)
     {
+        // protected states..
+        if (in_array($var, ['settings'])) {
+            return;
+        }
+
         $this->session->delete($var);
     }
 
@@ -103,6 +133,32 @@ abstract class Base implements Handler
         return $this->session->get($var, $default);
     }
 
+    /**
+     * @param array $settings
+     */
+    public function setSettings($settings)
+    {
+        $this->session->set('settings', $settings);
+    }
+
+    /**
+     * @param string $setting
+     * @param mixed $default
+     * @return mixed
+     */
+    public function getSetting($setting, $default = null)
+    {
+        $settings = $this->getState('settings', []);
+        if (isset($settings[$setting])) {
+            return $settings[$setting];
+        } else {
+            return null;
+        }
+    }
+
+    /**
+     * Utility to force the next answer to be sent to this handler
+     */
     protected function setWaitingAnswer()
     {
         $class = $this->getClassName();
